+++ workspace/widgets/metadata/vite.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'vite';
 
 import vue from '@vitejs/plugin-vue';
-import copy from "rollup-plugin-copy";
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -20,10 +19,7 @@ export default defineConfig({
                     entryFileNames: "assets/[name]-[hash].js",
                 },
             ],
-            plugins: [
-                copy({
-                    verbose: true,
-                }),
+            plugins: [,
             ],
         },
     },